feat: clean up using audit
